+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 0.1.157
+
+### Patch Changes
+
+- fix popover,menu
+  - @jamsr-ui/react@0.0.175
+
 ## 0.1.156
 
 ### Patch Changes